+++ ankihub/entry_point.py
@@ -33,7 +33,7 @@ from .gui.menu import menu_state, refresh_ankihub_menu, setup_ankihub_menu
 from .gui.operations.ankihub_sync import setup_full_sync_patch
 from .gui.operations.user_details import fetch_user_details_in_background
 from .gui.optimize_fsrs_dialog import maybe_show_fsrs_optimization_reminder
-from .main.block_exam_subdecks import check_and_handle_block_exam_subdeck_due_dates
+from .gui.subdeck_due_date_dialog import check_and_handle_block_exam_subdeck_due_dates
 from .main.note_deletion import handle_notes_deleted_from_webapp
 from .main.utils import modify_note_type_templates
 from .settings import (
