+++ tests/addon/test_integration.py
@@ -8930,59 +8930,3 @@ class TestCheckBlockExamSubdeckDueDates:
             assert "deck1" in expired_deck_ids  # past date
             assert "deck2" in expired_deck_ids  # today's date
             assert "deck3" not in expired_deck_ids  # future date
-
-    def test_check_block_exam_subdeck_due_dates_with_invalid_dates(
-        self,
-        anki_session_with_addon_data: AnkiSession,
-    ):
-        """Test function handles invalid date formats gracefully."""
-        with anki_session_with_addon_data.profile_loaded():
-            # Create configurations with invalid dates
-            config_items = [
-                BlockExamSubdeckConfig(ankihub_deck_id="deck1", subdeck_id="subdeck1", due_date="invalid-date"),
-                BlockExamSubdeckConfig(
-                    ankihub_deck_id="deck2",
-                    subdeck_id="subdeck2",
-                    due_date="2023-13-01",  # Invalid month
-                ),
-                BlockExamSubdeckConfig(
-                    ankihub_deck_id="deck3",
-                    subdeck_id="subdeck3",
-                    due_date="",  # Empty string
-                ),
-            ]
-
-            # Add configurations
-            for config_item in config_items:
-                config.add_block_exam_subdeck(config_item)
-
-            expired = check_block_exam_subdeck_due_dates()
-
-            # Should return empty list since all dates are invalid
-            assert expired == []
-
-    def test_check_block_exam_subdeck_due_dates_mixed_valid_invalid(
-        self,
-        anki_session_with_addon_data: AnkiSession,
-    ):
-        """Test function handles mix of valid and invalid dates."""
-        with anki_session_with_addon_data.profile_loaded():
-            # Mix of valid expired, valid future, and invalid dates
-            past_date = (date.today() - timedelta(days=1)).strftime("%Y-%m-%d")
-            future_date = (date.today() + timedelta(days=1)).strftime("%Y-%m-%d")
-
-            config_items = [
-                BlockExamSubdeckConfig(ankihub_deck_id="valid_expired", subdeck_id="subdeck1", due_date=past_date),
-                BlockExamSubdeckConfig(ankihub_deck_id="valid_future", subdeck_id="subdeck2", due_date=future_date),
-                BlockExamSubdeckConfig(ankihub_deck_id="invalid_date", subdeck_id="subdeck3", due_date="not-a-date"),
-            ]
-
-            # Add configurations
-            for config_item in config_items:
-                config.add_block_exam_subdeck(config_item)
-
-            expired = check_block_exam_subdeck_due_dates()
-
-            # Should return only the valid expired subdeck
-            assert len(expired) == 1
-            assert expired[0].ankihub_deck_id == "valid_expired"
